+++ apps/client/src/plugins/LayerSwitcher/LayerSwitcherUtils.js
